+++ hammer-sqldb/change-log.md
@@ -5,6 +5,7 @@
 3. 实现upsert在全部是插入时为entity设置自动生成的主键（存在update时返回的key无法确定）
 4. 优化GetOperate的ForUpdate逻辑
 5. Jdbc实现查询返回Tuple2,List<Tuple2>,Tuple3,List<Tuple3>,Tuple4,List<Tuple4>,Tuple5,List<Tuple5>
+6. SqldbHammer加入QueryEntity query(Table table);Update update(Table table);Delete delete(Table table)方法
 
 # 0.6.6 2022-08-26
 1. 修改L group(Consumer<C>)为L group(Function<C,L>)，